Broke out SidebarController class from MainActivity.java
